+++ IOCP/Acceptor.h
@@ -16,11 +16,11 @@
 */	
 #ifndef _ACCEPTOR_H
 #define _ACCEPTOR_H
-
+#include "stdint.h"
 typedef struct acceptor* acceptor_t;
 typedef void (*on_accept)(SOCKET,void*ud);
-acceptor_t create_acceptor(const char *ip,unsigned long port,on_accept accept_callback,void*ud);
+acceptor_t create_acceptor(const char *ip,uint32_t port,on_accept accept_callback,void*ud);
 void       destroy_acceptor(acceptor_t*);
-void       acceptor_run(acceptor_t,int ms);
+void       acceptor_run(acceptor_t,int32_t ms);
 
 #endif
